First attempt thread implementation
Appears slower when multithreading, with threading library.
"Notice that the user and sys both approximately sum to the real time. This is indicative that we gained no benefit from using the Threading library. If we had then we would expect the real time to be significantly less. These concepts within concurrent programming are usually known as CPU-time and wall-clock time respectively." taken from internet